Updated search indexer with user info queried from base chain.

+++ FreeMarketOne.Search/SellerAggregate.cs
@@ -7,7 +7,7 @@ namespace FreeMarketOne.Search
     public class SellerAggregate
     {
         public double TotalXRCVolume { get; set; }
-        public int StarRating { get; set; }
+        public double StarRating { get; set; }
         public List<string> PublicKeyHashes { get; set; }
         public string SellerName { get; set; }
         public List<byte[]> PublicKeys { get; set; }
